Create README.md
We build time series of directed event networks for a set of spatial distances, and based on scan-statistics, the spatial distance that generates the strongest change of event network connections is chosen.   In addition, we propose an empirical random network event generator (ERGEN) to detect significant motifs throughout time. This generator preserves the spatial configuration but randomizes the order of the occurrence of events. To prevent that the large number of links from masking the count of motifs, we propose using standardized counts of motifs at each time slot. When time slots are not known based on the specific problem, several exploratory statistical tools such as multiple correspondence analysis (MCA) can be used to define them. Our  methodology is able to detect interaction radius in space, build time series of networks, and describe changes in its topology over time, by means of identification of different types of motifs that allows for the understanding of the spatiotemporal dynamics of the phenomena. We illustrate our methodology by analyzing the dynamics of thefts in Medell\'in, (Colombia) between the years 2003 and 2015.